NMS-13775: Use hibernate type for bytea

+++ features/device-config/persistence/api/src/main/java/org/opennms/features/deviceconfig/persistence/api/DeviceConfigDao.java
@@ -30,5 +30,5 @@ package org.opennms.features.deviceconfig.persistence.api;
 
 import org.opennms.netmgt.dao.api.OnmsDao;
 
-public interface DeviceConfigDao extends OnmsDao<DeviceConfig, Integer> {
+public interface DeviceConfigDao extends OnmsDao<DeviceConfig, Long> {
 }
